+++ src/uniref/components/entry/SequenceSection.tsx
@@ -1,10 +1,9 @@
 import { Card, LongNumber, Sequence } from 'franklin-sites';
 
 import { hasContent } from '../../../shared/utils/utils';
+import { getEntrySectionNameAndId } from '../../utils/entrySection';
 
-import EntrySection, {
-  getEntrySectionNameAndId,
-} from '../../types/entrySection';
+import EntrySection from '../../types/entrySection';
 import { Sequence as SequenceType } from '../../../shared/types/sequence';
 
 const SequenceSection = ({ data }: { data: SequenceType }) => {
