+++ src/uniref/types/entrySection.ts
@@ -1,4 +1,4 @@
-export enum EntrySection {
+enum EntrySection {
   Sequence = 'sequence',
   Members = 'members',
 }
@@ -6,25 +6,6 @@ export enum EntrySection {
 export type EntrySectionNameAndId = {
   name: string;
   id: EntrySection;
-};
-
-export const getEntrySectionNameAndId = (
-  section: EntrySection
-  // eslint-disable-next-line consistent-return
-): EntrySectionNameAndId => {
-  // eslint-disable-next-line default-case
-  switch (section) {
-    case EntrySection.Sequence:
-      return {
-        name: 'Representative',
-        id: EntrySection.Sequence,
-      };
-    case EntrySection.Members:
-      return {
-        name: 'member',
-        id: EntrySection.Members,
-      };
-  }
 };
 
 export default EntrySection;
